feat: Renamed folder to core-sdk-v10
