+++ README.md
@@ -19,6 +19,12 @@ Para visualizar a documentação de cada módulo, utilize os links abaixo:
 - [Guia rápido de atualização](https://github.com/oititec/liveness-react-example/blob/main/update-guide/README.
 md)
 
+### [FaceTec SDK: 10.0.42](https://github.com/oititec/liveness-react-example/main/core/core-sdk-v10)
+
+_data: 29/04/2026_
+
+- Nova implementação de liveness 3D com o SDK web da Facetec versão 10.0.42
+
 ### [Iproov SDK: 6.0.2](https://github.com/oititec/liveness-react-example/blob/main/src/app/iproov/README.md)
 
 _data: 11/03/2026_
