+++ core/core-sdk-v10/core-sdk-optional/FaceTecStrings.el.js
