+++ core/core-sdk-v10/core-sdk-optional/FaceTecStrings.ru.js
