+++ core/core-sdk-v10/core-sdk/FaceTecSDK.js/FaceTecLogging.d.ts
