+++ core/core-sdk-v10/core-sdk/FaceTecSDK.js/FaceTecPublicApi.d.ts
