+++ core/core-sdk-v10/core-sdk/FaceTecSDK.js/FaceTecSession.d.ts
