+++ core/core-sdk-v10/core-sdk/FaceTecSDK.js/resources/011c90516755d702cfb4205ca9d93e21fe6683b8.wasm
