+++ core/core-sdk-v10/core-sdk/FaceTecSDK.js/resources/Faced35693a775734e6fa9bf6ccdaf00a25f.js
