+++ core/core-sdk-v10/core-sdk/FaceTec_images/FaceTec_rotate.png
