+++ core/core-sdk-v10/core-sdk/FaceTec_images/FaceTec_your_app_logo.png
